fix: extract isAllowedNavigation to own module to fix Windows CI
The test imported from ./index which pulls in all Electron imports,
causing failures on Windows CI where Electron is not available.
Extract isAllowedNavigation to navigation-guard.ts so tests can
import it without loading the full Electron main process.
